chore: load data from page, not from tab
